refactor: fix folder name in printing
