+++ packages/web/src/features/printing-business/components/PrintingBusinessForm/PrintingBusinessFormThird.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import Card from "@sparcs-clubs/web/common/components/Card";
 import TextInput from "@sparcs-clubs/web/common/components/Forms/TextInput";
 import { printingBusinessOrderSteps } from "@sparcs-clubs/web/constants/printingBusiness";
-import DesiredPickUpTimeSelection from "@sparcs-clubs/web/features/printing-business/component/DesiredPickUpTimeSelection/DesiredPickUpTimeSelection";
+import DesiredPickUpTimeSelection from "@sparcs-clubs/web/features/printing-business/components/DesiredPickUpTimeSelection/DesiredPickUpTimeSelection";
 import { mockExistDates } from "@sparcs-clubs/web/features/rental-business/services/_mock/mockExistDate";
 
 import type { PrintingBusinessFormProps } from ".";
